More strictly hint URL pattern members (#218)
* More strictly hint URL pattern members

In many cases `Sequence[URLPattern]` might be sufficient -- as long as the pattern doesn't include `include()`. But determining where include patterns can or cannot occur is difficult.

Using `Sequence[Union...]` instead of `List[Union...]`, because the former is covariant. Users may be passing an auto-inferred list of patterns, which may be `List[URLPattern]`

* Alias _AnyURL = Union[URLPattern, URLResolver]

* Fix one more Sequence[_AnyURL] -> Iterable