Fix Renovate 'lock file maintenance' schedule *again*
Sorry, again!

Turns out that the interaction between Renovate Cloud scheduling and `"schedule"` setting is confusing and poorly documented.

`lockFileMaintenance` has a different default schedule -- it defaults to `before 4am on monday`. Once per week is too rare IMO for our use case -- and the "before 4am" part is rarely useful with the cloud offering, because the triggering time is arbitrary during the day.